% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea-utils.R
\name{calculate_scale_eff}
\alias{calculate_scale_eff}
\title{Calculate scale efficiency}
\usage{
calculate_scale_eff(
  inputs,
  outputs,
  orientation = c("in", "out"),
  digits = NULL
)
}
\arguments{
\item{inputs}{A matrix of inputs, created with \code{create_matrix}}

\item{outputs}{A matrix of outputs, created with \code{create_matrix}}

\item{orientation}{\verb{in} for input oriented models or \code{out} for output oriented}

\item{digits}{An integer with the number of digits to round to. If \code{NULL} the
values are kept unrounded.}
}
\description{
Calculate scale efficiency from a set of inputs and outputs and return a
data.frame
}
