% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-dea.R
\name{compute_super_efficiency}
\alias{compute_super_efficiency}
\title{Compute super efficiency}
\usage{
compute_super_efficiency(
  x,
  y,
  type = c("crs", "vrs", "drs", "irs"),
  orientation = c("in", "out"),
  digits = NULL
)
}
\arguments{
\item{x}{A matrix with input values.}

\item{y}{A matrix with output values.}

\item{type}{Type of model.
\tabular{rlll}{
\tab vrs \tab\tab Variable returns to scale, convexity and free disposability. \cr
\tab drs \tab\tab Decreasing returns to scale, convexity, down-scaling and free disposability. \cr
\tab crs \tab\tab Constant returns to scale, convexity and free disposability. \cr
\tab irs \tab\tab Increasing returns to scale, (up-scaling, but not down-scaling), convexity and free disposability.
}}

\item{orientation}{Model orientation.}

\item{digits}{Number of decimal digits in the response. Defaults to \code{NULL} (no rounding).}
}
\description{
Calculate DEA super efficiency scores.
}
