% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea-utils.R
\name{create_matrix}
\alias{create_matrix}
\title{Create a matrix for input or output variables}
\usage{
create_matrix(df, columns, id, normalize = FALSE)
}
\arguments{
\item{df}{A data.frame}

\item{columns}{A vector of column names that should be included in the matrix}

\item{id}{The name of the column with the DMU IDs}

\item{normalize}{If \code{TRUE}, all columns will be normalized with a mean of 1}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Create a matrix for input or output variables that can be used in DEA models
from a supplied data.frame
}
