% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-dea.R
\name{get_peers}
\alias{get_peers}
\title{Get peers}
\usage{
get_peers(lambda, ids, threshold = 0)
}
\arguments{
\item{lambda}{Output of \code{compute_efficiency()$lambda}.}

\item{ids}{A vector with DMU names or ids.}

\item{threshold}{Minimum weight for extracted peers. Defaults to 0.}
}
\value{
data.frame
}
\description{
Find peers for each DMU.
}
