% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{run_pioneer}
\alias{run_pioneer}
\alias{runPioneeR}
\title{Run pioneeR}
\usage{
run_pioneer(x = NULL, port = NULL, ...)

runPioneeR(x = NULL, port = NULL, ...)
}
\arguments{
\item{x}{A data frame that should be loaded with the app. See details.}

\item{port}{The TCP port that the application should listen on.}

\item{...}{Other arguments to send to \code{\link[shiny]{runApp}}}
}
\description{
Run the pioneeR app on your local machine.
}
\details{
You can load a data object in your current environment to the app. You can pass
a data frame or a character string with the object name of the data frame you
want to be loaded when the app launches. Note that you should only use data
frame objects. If you have a tibble (from the tidyverse) or a data table, you
can convert to an ordinary data.frame using \code{\link[base]{as.data.frame}}.
}
