% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa.R
\name{sfa}
\alias{sfa}
\title{Stochastic Frontier Analysis}
\usage{
sfa(formula, data = NULL, pars = NULL, productivity = TRUE)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the fixed-effects part of the model, with the response on the left of a ~ operator and the terms, separated by + operators, on the right.}

\item{data}{a data frame containing the variables named in formula. By default the variables are taken from the environment from which sfa is called.}

\item{pars}{Starting values for the algorithm.}

\item{productivity}{Boolean indicating productivity (TRUE) or cost analysis (FALSE).}
}
\value{
\item{coef}{a vector of coefficients of the model. The coefficients start with the intercept, followed by the regression coefficients of the variables. The vector ends with the error variance, and the variance of the half normal distribution}

\item{hess}{the hessian of the fitted model}

\item{loglik}{the optimized log likelihood}

\item{ols}{regression coefficients for the starting values}

\item{sigmau2}{variance of the half normal distribution}

\item{sigmav2}{variance of the errors}

\item{sigma2}{The estimate of the total variance, that is, the sum of two sigma's}

\item{gamma}{Ratio of the variance of the half normal distribution over the total variance}

\item{convergence}{If the value is zero the model converged, other values indicate problems during iterations}

\item{ptable}{summary table with coefficients, standard errors, z values, and p values}
}
\description{
This function computes a stochastic frontier model using the half normal model for the efficiencies.
The fitting method is Maximum Likelihood and both Production and Cost Functions can be estimated.
}
