% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-dea.R
\name{summary_tbl_dea}
\alias{summary_tbl_dea}
\title{Create a summary table for DEA}
\usage{
summary_tbl_dea(x)
}
\arguments{
\item{x}{A vector of efficiency scores or a Farrell object}
}
\value{
data.frame
}
\description{
Create a binned summary table for efficiency scores from a DEA model.
}
\details{
The function will return a summary table for efficiency scores from a DEA model.
Efficiency scores will be placed in 11 bins, where DMUs with an effciency score
equal to 1 are placed in a separate bin. For output oriented models with range
[1, Inf], bins are created with \code{1/bin}. Bin widths will be equal to models
with range [0, 1].
}
