% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{unset_env_vars}
\alias{unset_env_vars}
\title{Unset environment variables}
\usage{
unset_env_vars()
}
\value{
A logical vector, with elements being \code{TRUE} if unsetting the variable succeeded
}
\description{
Unsets the environment variables set by pioneeR
}
\examples{
unset_env_vars()
}
